✨ Remote functionsに一部置き換え
