+++ package.json
@@ -16,12 +16,13 @@
 		"autoprefixer": "^10.4.21",
 		"cssnano": "^7.1.2",
 		"eslint": "^9.38.0",
-		"eslint-plugin-svelte": "^3.13.0"
+		"eslint-plugin-svelte": "^3.13.0",
+		"valibot": "^1.2.0"
 	},
 	"type": "module",
 	"dependencies": {
 		"@sveltejs/adapter-cloudflare": "^7.2.4",
-		"@sveltejs/kit": "^2.48.3",
+		"@sveltejs/kit": "^2.49.2",
 		"@typescript-eslint/eslint-plugin": "^8.46.2",
 		"@typescript-eslint/parser": "^8.46.2",
 		"dayjs": "^1.11.18",
@@ -30,12 +31,12 @@
 		"postcss": "^8.5.6",
 		"prettier": "^3.6.2",
 		"prettier-plugin-svelte": "^3.4.0",
-		"svelte": "^5.43.0",
-		"svelte-check": "^4.3.3",
+		"svelte": "^5.45.8",
+		"svelte-check": "^4.3.4",
 		"tailwindcss": "^3.4.17",
 		"typescript": "^5.9.3",
-		"vite": "npm:rolldown-vite@^7.1.12",
-		"wrangler": "^4.45.2"
+		"vite": "npm:rolldown-vite@^7.2.7",
+		"wrangler": "^4.53.0"
 	},
 	"pnpm": {
 		"overrides": {
