--- src/lib/schema/tag/tag.ts
@@ -1,5 +1,0 @@
-import type { MicroCMSListContent } from '$lib/libs/microcms.server';
-export type Tag = {
-	title: string;
-	slug: string;
-} & MicroCMSListContent;
