--- src/lib/schema/thinking/contents.ts
@@ -1,3 +1,0 @@
-export type Contents = {
-	body: string;
-};
