--- src/lib/schema/thinking/meta.ts
@@ -1,4 +1,0 @@
-export type Meta = {
-	title?: string;
-	description?: string;
-};
