--- src/lib/schema/thinking/section.ts
@@ -1,12 +1,0 @@
-import type { Contents } from './contents';
-export type Section = {
-	title: string;
-	contents: (
-		| ({
-				fieldId: 'contents';
-		  } & Contents)
-		| ({
-				fieldId: 'section';
-		  } & Section)
-	)[];
-};
