+++ src/routes/[slug]/+page.server.ts
@@ -1,8 +1,8 @@
 import { error } from '@sveltejs/kit';
-import type { PageServerLoad } from './$types';
 import { client } from '$lib/libs/microcms.server';
 import type { StaticPageResponse } from '$lib/schema/staticPage/staticPage';
-export const load = (async ({ params }) => {
+
+export const load = async ({ params }) => {
 	const pages = await client.getList<StaticPageResponse>({
 		endpoint: 'static_page',
 		queries: {
@@ -17,4 +17,4 @@ export const load = (async ({ params }) => {
 		};
 	}
 	error(400, '記事を取得できませんでした。');
-}) satisfies PageServerLoad;
+};
